restructured the plot and report generation in separate scripts, to be tested
